[run ci] fixed the merge issues
